refactor: move keyhint functionality out of the guimanager
