+++ JotunnLib/Main.cs
@@ -61,6 +61,7 @@ namespace Jotunn
                 MockManager.Instance,
                 KitbashManager.Instance,
                 GUIManager.Instance,
+                KeyHintManager.Instance,
                 //SaveManager.Instance,  // Temporarely disabled, causes FPS issues in the current implementation
                 //ZoneManager.Instance,  // Had some problems reported, needs more tests
                 SynchronizationManager.Instance
